+++ plugins/infoproviders/infoprovider.go
@@ -1,0 +1,5 @@
+package infoproviders
+
+type InfoProvider interface {
+	Get() interface{}
+}
